Final Docker setup

+++ transaction-service/src/test/java/ericphamm/transaction_service/controller/TransactionControllerTest.java
@@ -11,6 +11,7 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
@@ -32,7 +33,7 @@ class TransactionControllerTest {
         Transaction mockTx = new Transaction();
         mockTx.setId(1L);
         mockTx.setDescription("Test");
-        mockTx.setAmount(100.0);
+        mockTx.setAmount(BigDecimal.valueOf(100.0));
 
         Mockito.when(transactionService.getAllSortedByTimestamp())
                 .thenReturn(List.of(mockTx));
@@ -49,7 +50,7 @@ class TransactionControllerTest {
         Transaction mockTx = new Transaction();
         mockTx.setId(1L);
         mockTx.setDescription("Created Tx");
-        mockTx.setAmount(123.45);
+        mockTx.setAmount(BigDecimal.valueOf(123.45));
 
         Mockito.when(transactionService.createTransaction(Mockito.any()))
                 .thenReturn(mockTx);
